테스트용1
2023.04.13

+++ jPanda/src/main/java/com/kakao/jPanda/lhw/domain/Notice.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class Notice {
 	// 공지사항 게시판
-	private int noticeNo;
+	private Long noticeNo;
 	private String title;
 	private String content;
 	private Date regDate;
